add adversarial bandit env + split bandit envs
