+++ rlberry/envs/bandits/__init__.py
@@ -1,1 +1,3 @@
-from .bandits import *
+from .bandit_base import AdversarialBandit, Bandit
+from .corrupted_bandits import *
+from .stochastic_bandits import *
